+++ proxy/Cargo.toml
@@ -32,6 +32,7 @@ thiserror = "1.0.30"
 tokio = { version = "1.17", features = ["macros"] }
 tokio-postgres = { git = "https://github.com/zenithdb/rust-postgres.git", rev="d052ee8b86fff9897c77b0fe89ea9daba0e1fa38" }
 tokio-rustls = "0.23.0"
+url = "2.2.2"
 
 utils = { path = "../libs/utils" }
 metrics = { path = "../libs/metrics" }
